+++ MaxMind.GeoIP2/Model/Postal.cs
@@ -1,34 +1,13 @@
-﻿#region
-
 using MaxMind.Db;
 using System.Text.Json.Serialization;
-
-#endregion
 
 namespace MaxMind.GeoIP2.Model
 {
     /// <summary>
     ///     Contains data for the postal record associated with an IP address.
     /// </summary>
-    public class Postal
+    public record Postal
     {
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        public Postal()
-        {
-        }
-
-        /// <summary>
-        ///     Constructor
-        /// </summary>
-        [Constructor]
-        public Postal(string? code = null, int? confidence = null)
-        {
-            Code = code;
-            Confidence = confidence;
-        }
-
         /// <summary>
         ///     The postal code of the location. Postal codes are not available
         ///     for all countries. In some countries, this will only contain part
@@ -36,7 +15,8 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("code")]
-        public string? Code { get; internal set; }
+        [MapKey("code")]
+        public string? Code { get; init; }
 
         /// <summary>
         ///     A value from 0-100 indicating MaxMind's confidence that the
@@ -45,17 +25,7 @@ namespace MaxMind.GeoIP2.Model
         /// </summary>
         [JsonInclude]
         [JsonPropertyName("confidence")]
-        public int? Confidence { get; internal set; }
-
-        /// <summary>
-        ///     Returns a <see cref="string" /> that represents this instance.
-        /// </summary>
-        /// <returns>
-        ///     A <see cref="string" /> that represents this instance.
-        /// </returns>
-        public override string ToString()
-        {
-            return $"Code: {Code}, Confidence: {Confidence}";
-        }
+        [MapKey("confidence")]
+        public int? Confidence { get; init; }
     }
 }